refactor(db) 3/13: db/schema.rs dispatcher + atomic create_tables (closes #249)
Move the 1,382-line DDL execute_batch out of init_schema into
schema::create_tables(tx). init_schema is now a dispatcher: all DDL in one
explicit transaction (atomic; direct tx not with_tx since with_tx would
flush into not-yet-created tables), then the two migrate_* after on a
fresh lock (ordering preserved). All CREATE ... IF NOT EXISTS (idempotent).
Added init_schema_is_idempotent_and_creates_all_tables test (71 tables).
240 tests + clippy -D warnings + fmt green.
